#799: lb: major cleanup and re-factor of StatsRestartReader
