+++ src/vt/runtime/runtime_get.cc
@@ -128,6 +128,7 @@ objgroup::ObjGroupManager*  theObjGroup()           { return CUR_RT->theObjGroup
 rdma::Manager*              theHandleRDMA()         { return CUR_RT->theHandleRDMA;     }
 util::memory::MemoryUsage*  theMemUsage()           { return CUR_RT->theMemUsage;       }
 vrt::collection::balance::ProcStats* theProcStats() { return CUR_RT->theProcStats;      }
+vrt::collection::balance::StatsRestartReader* theStatsReader() { return CUR_RT->theStatsReader;      }
 
 #if backend_check_enabled(trace_enabled)
 trace::Trace*               theTrace()          { return CUR_RT->theTrace;          }
